feat: Add function to return string as a char ptr
This function returns the internal char pointer used by the StrBuilder
object. Note that it is not guaranteed to be NUL terminated, as the API
uses the length property to keep track of the string in memory.

The function "strbuilder_to_cstr" allocates a new char buffer with the
size of length + 1 and copies the string into it, adding a NUL character
at the end.