feat: keep last error in StrBuilder object
To help keep track of errors, the StrBuilder object now holds the last registered error. This value can be obtained by calling strbuilder_get_error or  strbuilder_get_error_msg.

Furthermore, the macro SET_ERROR_RETURN was added with the objective of compressing the operation of assigning the error value and then returning it into one line.

Lastly, I replaced calloc with malloc in strbuilder_create_sz as the memory initialization is not really required in this scenario.