feat(refcount): memory management using refcounting
This is an experimental code that uses a reference counting technique to manage the memory used by the API. The refcount keeps track of how many StrBuilder instances reference the same string in memory. The API uses a copy-on-write algorithm, where strings are copied only when they need to be modified and their reference count is greater than one.
When the refcount is equal to 1, this means that the string is only referenced by the StrBuilder object that contains it; therefore, it is not necessary to make a copy and such a block of memory can be modified or reallocated. Any function, that doesn't modify the contents of the memory, may increase the refcounter by one. Once the refcount reaches 0, the string (and its memory) will be freed during the execution of strbuilder_free call.

The general behavior of the API had to be modified slightly: now all functions that modify the string, add a null terminator automatically.